Partition scraper logic into base and legacy files
